Add files via upload
A senior salesperson is paid $800 a week and a junior salesperson $375 a week. Write a program that accepts as input a salesperson’s status in the character variable. If the status entered is’, the senior person’s salary will be displayed and else the junior person’s salary should be output. A sample run would be as follows:

Sample Run: 
Please enter the status of the salesperson (s/j): s 
The salary of the senior salesperson is $800 per week.